Add a attribution reference to the Moo ICT code
Legally, this isn't strictly necesary, but it's good practice.
(And if Moo ICT was Apache licensed it would be necessary.)

If there are any other files based on the Moo ICT code, it'd be
good to add the comment at the top of them too.

It's also useful for folks that want to look at the original code,
for referenence.